[#297] refactor(recommendation): 신청자 추천 컨텍스트 조회 로직 분리
